chore: fix circular import issue
